fixed legacy build
